+++ Keller_Pressure_Buffer.h
@@ -16,11 +16,11 @@
 typedef struct {
     int32_t p_mbar;
     int32_t t_centi;
-    uint32_t t_ticks;
+    uint64_t t_ticks;
 } keller_sample_t;
 
 void keller_buffer_init(void);
-bool keller_buffer_store(int32_t p_mbar, int32_t t_centi, uint32_t t_ticks);
+bool keller_buffer_store(int32_t p_mbar, int32_t t_centi, uint64_t t_ticks);
 bool keller_buffer_retrieve(keller_sample_t *sample);
 bool keller_buffer_is_empty(void);
